cloud-file/sl-prediction-cloud-function/requirements.txt
# Function dependencies, for example:
# package>=version

numpy>=1.18.1
tensorflow
google-cloud-storage>=1.38.0
opencv-python
firebase-admin